linux-yocto: Add a user config fragment for rapid development
To easily modify the kernel configuration (and make it stick),
just add the CONFIG_* options you need to the new user.cfg file,
and rebuild the kernel.

The initial version of the user.cfg file adds support for USB Ethernet.
For now, I suggest sticking to built-ins as it makes things a bit less
complex for development.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Cc: Koen Kooi <koen@circuitco.com>