xserver-xf86-config: make overrides work with a large layer stack
When one of the BSP layers in the stack uses := with THISDIR all other
BSPs bbappends for this recipe need to use it as well to avoid pointing
it to the wrong thing.

This fixes the bug where the fallback xorg.conf gets included and not
the proper meta-minnow EMGD one for minnow when building this recipe
with the full angstrom layer stack.

Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>
Signed-off-by: Darren Hart <dvhart@linux.intel.com>